backend/requirements.txt
# =========================
# Django Core
# =========================
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3

# =========================
# Database
# =========================
psycopg2-binary>=2.9
dj-database-url>=2.1
django-cockroachdb>=5.0

# =========================
# Authentication
# =========================
djangorestframework-simplejwt>=5.3

# =========================
# Utilities
# =========================
django-filter>=23.5
Pillow>=10.0
python-decouple>=3.8

# =========================
# Admin UI
# =========================
django-jazzmin>=2.6

# =========================
# Production Server
# =========================
gunicorn>=21.2
whitenoise>=6.6

# =========================
# Development & Testing
# =========================
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3

# =========================
# Code Quality
# =========================
flake8>=6.1
black>=23.12
isort>=5.13
